DefaultViewFinder is no longer a ViewFinder
